refactor components out of containers
